+++ GameEngine/src/main/java/au/edu/newcastle/SENG48002013/game/engine/model/actions/InputDirectionAction.java
@@ -36,7 +36,7 @@ public class InputDirectionAction extends BaseInputAction {
 		this.gameObject = gameObject;
 	}
 
-	public int doAction(long dt)
+	public int doAction(double dt)
 	{
 		if(getInput() != null)
 		{
