+++ GameEngine/src/main/java/au/edu/newcastle/SENG48002013/game/engine/model/actions/InputPosAction.java
@@ -23,7 +23,7 @@ public class InputPosAction extends BaseInputAction {
 		this.gameObject = gameObject;
 	}
 
-	public int doAction(long dt)
+	public int doAction(double dt)
 	{
 		if(getInput() != null)
 		{
