+++ GameEngine/src/main/java/au/edu/newcastle/SENG48002013/game/engine/model/actions/InputVelAction.java
@@ -22,7 +22,7 @@ public class InputVelAction extends BaseInputAction {
 	{
 		this.gameObject = gameObject;
 	}
-	public int doAction(long dt)
+	public int doAction(double dt)
 	{
 		if(getInput() != null)
 		{
