+++ GameEngine/src/main/java/au/edu/newcastle/SENG48002013/game/engine/model/events/TimerEvent.java
@@ -38,7 +38,7 @@ public class TimerEvent extends BaseEvent {
 		this.repeat = repeat;
 	}
 	@Override
-	public int evaluate(long dt)
+	public int evaluate(double dt)
 	{
 		int returnCode = -1;
 		currentTime -= dt;
